implement CRUD methods

+++ src/main/java/com/exalt/sampleproject/service/RestaurantsService.java
@@ -1,12 +1,10 @@
 package com.exalt.sampleproject.service;
 
-import com.exalt.sampleproject.model.Contact;
-import com.exalt.sampleproject.model.Location;
+import com.exalt.sampleproject.dto.ResponseMessage;
 import com.exalt.sampleproject.model.Restaurants;
 import com.exalt.sampleproject.repository.RestaurantsRepo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,36 +13,38 @@ import java.util.Optional;
 @Service
 public class RestaurantsService {
     private final RestaurantsRepo restaurantsRepo;
-    private final LocationService locationService;
-    private final ContactService contactService;
 
-    Logger logger = LoggerFactory.getLogger(RestaurantsService.class);
 
-    public RestaurantsService(RestaurantsRepo restaurantsRepo, LocationService locationService, ContactService contactService) {
+    private final static Logger logger = LoggerFactory.getLogger(RestaurantsService.class);
+
+    public RestaurantsService(RestaurantsRepo restaurantsRepo) {
         this.restaurantsRepo = restaurantsRepo;
-        this.locationService = locationService;
-        this.contactService = contactService;
     }
 
+//
+//    public void createRecord() {
+//        Restaurants res = new Restaurants("KAN YA MAKAN");
+//        Restaurants res2 = new Restaurants("ward");
+//
+//        Location location = new Location("nablus", "rafidya", "akadimya", res);
+//        Contact contact = new Contact("092345125", "0597430457", res);
+//
+//        res.setContact(contact);
+//        res.addLocation(location);
+//
+//       // logger.info("location: " + res.getLocation().getCity() + " " + res.getLocation().getSection());
+//        save(res);
+//        save(res2);
+//    }
 
-    @Autowired
-    public void createRecord() {
-        Restaurants res = new Restaurants("KAN YA MAKAN");
-        Restaurants res2 = new Restaurants("ward");
+    public ResponseMessage save(Restaurants restaurants) {
+        restaurantsRepo.save(restaurants);
 
-        Location location = new Location("nablus", "rafidya", "akadimya", res);
-        Contact contact = new Contact("092345125", "0597430457", res);
-        locationService.save(location);
-        contactService.save(contact);
+        ResponseMessage responseMessage = new ResponseMessage();
+        responseMessage.setMessage("Successfully saved");
+        responseMessage.setStatus(1);
 
-        res.addContact(contact);
-        res.addLocation(location);
-        save(res);
-        save(res2);
-    }
-
-    public Restaurants save(Restaurants restaurants) {
-        return restaurantsRepo.save(restaurants);
+        return responseMessage;
     }
 
     public List<Restaurants> findAll() {
@@ -55,21 +55,39 @@ public class RestaurantsService {
         return restaurantsRepo.findById(id);
     }
 
-    public Restaurants updateRestaurant(Long id, Restaurants updatedRestaurant) {
+    public ResponseMessage updateRestaurant(Long id, Restaurants updatedRestaurant) {
+        ResponseMessage responseMessage = new ResponseMessage();
+        Optional<Restaurants> optionalRestaurants = findById(id);
 
-        return findById(id)
-                .map(restaurant -> {
-                    restaurant.setName(updatedRestaurant.getName());
-                    restaurant.setContacts(updatedRestaurant.getContacts());
-                    restaurant.setLocations(updatedRestaurant.getLocations());
-                    return restaurantsRepo.save(restaurant);
-                })
-                .orElseGet(() -> {
-                    return restaurantsRepo.save(updatedRestaurant);
-                });
+        if (optionalRestaurants.isPresent()) {
+            optionalRestaurants.map(restaurant -> {
+                restaurant.setName(updatedRestaurant.getName());
+                restaurant.setContact(updatedRestaurant.getContact());
+                restaurant.setLocations(updatedRestaurant.getLocations());
+                save(restaurant);
+                return restaurant;
+            });
+            responseMessage.setMessage("Success updating Restaurant");
+            responseMessage.setStatus(1);
+        } else {
+            responseMessage.setMessage("RestaurantId " + id + ", was not found.");
+            responseMessage.setStatus(-1);
+        }
+        return responseMessage;
     }
 
-    public void deleteRestaurant(Long id) {
-        restaurantsRepo.delete(findById(id).get());
+    public ResponseMessage deleteRestaurant(Long id) {
+        ResponseMessage responseMessage = new ResponseMessage();
+
+        if (findById(id).isPresent()) {
+            restaurantsRepo.delete(findById(id).get());
+            responseMessage.setMessage("Successfully deleted");
+            responseMessage.setStatus(1);
+        } else {
+            responseMessage.setMessage("Not found");
+            responseMessage.setStatus(-1);
+        }
+
+        return responseMessage;
     }
 }
